extract utils into own file
